+++ agent/lm_agent/reconciliation.py
@@ -19,7 +19,7 @@ from lm_agent.license_report import report
 from lm_agent.logs import logger
 from lm_agent.server_interfaces.license_server_interface import LicenseReportItem
 from lm_agent.workload_managers.slurm.cmd_utils import (
-    get_all_features_used_values,
+    get_all_features_cluster_values,
     return_formatted_squeue_out,
     squeue_parser,
 )
@@ -135,7 +135,7 @@ async def reconcile():
     all_features_bookings_sum = await get_all_features_bookings_sum()
 
     # Get license usage from the cluster
-    all_features_used_value = await get_all_features_used_values()
+    all_features_cluster_value = await get_all_features_cluster_values()
 
     # Delete bookings for jobs that reached the grace time
     logger.debug("Cleaning jobs by grace time")
@@ -153,8 +153,8 @@ async def reconcile():
     for license_data in license_usage_info:
         # Get license usage from license report
         product_feature = license_data.product_feature
-        server_used = license_data.used
-        total = license_data.total
+        report_used = license_data.used
+        report_total = license_data.total
 
         # Get booking information from backend
         booking_sum = all_features_bookings_sum[product_feature]
@@ -167,23 +167,32 @@ async def reconcile():
                     reserved = feature.reserved
 
         # Get license usage from the cluster
-        slurm_used = all_features_used_value[product_feature]
+        slurm_used = all_features_cluster_value[product_feature]["used"]
+        slurm_total = all_features_cluster_value[product_feature]["total"]
 
         """
         The reserved amount represents how many licenses are already in use:
         Either in the license server or booked for a job (bookings from other cluster as well).
-        If the license has a reserved value, it should be reserved too.
 
-        The reservation is not meant to be used by any user, it's a way to block usage of licenses
+        If the license has a reserved value (licenses exclusive for usage in desktop environments),
+        it should be added to the reservation too.
+
+        The reservation is not meant to be used by any user, it's a way to block usage of licenses.
+
+        If the report total is 0, it means that the license is not available in the license server
+        and it should be fully reserved to prevent jobs from running and crashing.
         """
 
-        reservation_amount = server_used - slurm_used + booking_sum + reserved
+        if report_total == 0:
+            reservation_amount = slurm_total
+        else:
+            reservation_amount = report_used - slurm_used + booking_sum + reserved
 
         if reservation_amount < 0:
             reservation_amount = 0
 
-        if reservation_amount > total:
-            reservation_amount = total
+        if reservation_amount > slurm_total:
+            reservation_amount = slurm_total
 
         if reservation_amount:
             reservation_data.append(f"{product_feature}@{license_server_type}:{reservation_amount}")