+++ agent/tests/test_license_report.py
@@ -408,3 +408,52 @@ def test_get_local_license_configurations():
     assert license_report.get_local_license_configurations(license_configurations, local_licenses) == [
         configuration_super
     ]
+
+
+@mark.asyncio
+@mock.patch("lm_agent.license_report.get_cluster_configs_from_backend")
+@mock.patch("lm_agent.license_report.get_local_license_configurations")
+@mock.patch("lm_agent.license_report.RLMLicenseServer.get_report_item")
+async def test_license_report_empty_on_exception_raised(
+    get_report_item_mock: mock.MagicMock,
+    get_local_license_configurations_mock: mock.MagicMock,
+    get_cluster_configs_from_backend_mock: mock.MagicMock,
+):
+    """
+    Do I get an empty report when an exception is raised?
+    """
+    get_cluster_configs_from_backend_mock.return_value = []
+    get_local_license_configurations_mock.return_value = [
+        ConfigurationSchema(
+            id=1,
+            name="Converge",
+            cluster_client_id="dummy",
+            features=[
+                FeatureSchema(
+                    id=1,
+                    name="converge_super",
+                    product=ProductSchema(id=1, name="converge"),
+                    config_id=1,
+                    reserved=100,
+                    total=1000,
+                    used=93,
+                    booked_total=0,
+                )
+            ],
+            license_servers=[
+                LicenseServerSchema(id=1, config_id=1, host="licserv0001", port=1234),
+                LicenseServerSchema(id=3, config_id=1, host="licserv0003", port=8760),
+            ],
+            grace_time=60,
+            type=LicenseServerType.RLM,
+        )
+    ]
+    get_report_item_mock.side_effect = Exception("Something is wrong with the license server!")
+
+    assert await license_report.report() == [
+        {
+            "product_feature": "converge.converge_super",
+            "used": 0,
+            "total": 0,
+        }
+    ]